requirements.txt
selenium
webdriver-manager
pillow
requests
html2text
argparse
markdown-it-py[plugins]
